# idle-backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Max Idle API",
    "version": "1.0.0",
    "description": "Generated API contract for Max Idle backend."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "TimeCurrencyBalances": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "available": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "total",
          "available"
        ]
      },
      "PlayerState": {
        "type": "object",
        "properties": {
          "idleTime": {
            "$ref": "#/components/schemas/TimeCurrencyBalances"
          },
          "realTime": {
            "$ref": "#/components/schemas/TimeCurrencyBalances"
          },
          "timeGems": {
            "$ref": "#/components/schemas/TimeCurrencyBalances"
          },
          "upgradesPurchased": {
            "type": "integer",
            "minimum": 0
          },
          "currentSeconds": {
            "type": "integer",
            "minimum": 0
          },
          "idleSecondsRate": {
            "type": "number",
            "minimum": 0
          },
          "secondsMultiplier": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "shop": {
            "type": "object",
            "properties": {
              "seconds_multiplier": {
                "type": "integer",
                "minimum": 0
              },
              "restraint": {
                "type": "integer",
                "minimum": 0
              },
              "idle_hoarder": {
                "type": "integer",
                "minimum": 0,
                "maximum": 5
              },
              "luck": {
                "type": "integer",
                "minimum": 0
              },
              "collect_gem_time_boost": {
                "type": "integer",
                "minimum": 0,
                "maximum": 5
              },
              "worthwhile_achievements": {
                "type": "integer",
                "minimum": 0,
                "maximum": 5
              }
            },
            "required": [
              "seconds_multiplier",
              "restraint",
              "luck"
            ],
            "additionalProperties": {
              "nullable": true
            }
          },
          "achievementCount": {
            "type": "integer",
            "minimum": 0
          },
          "achievementBonusMultiplier": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "hasUnseenAchievements": {
            "type": "boolean"
          },
          "currentSecondsLastUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "lastCollectedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastDailyRewardCollectedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "idleTime",
          "realTime",
          "timeGems",
          "upgradesPurchased",
          "currentSeconds",
          "idleSecondsRate",
          "secondsMultiplier",
          "shop",
          "achievementCount",
          "achievementBonusMultiplier",
          "hasUnseenAchievements",
          "currentSecondsLastUpdated",
          "lastCollectedAt",
          "lastDailyRewardCollectedAt",
          "serverTime"
        ]
      },
...
      "PlayerCollectResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PlayerState"
          },
          {
            "type": "object",
            "properties": {
              "collectedSeconds": {
                "type": "integer",
                "minimum": 0
              },
              "realSecondsCollected": {
                "type": "integer",
                "minimum": 0
              }
            },
            "required": [
              "collectedSeconds",
              "realSecondsCollected"
            ]
          }
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/player": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get current player state",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current player state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerState"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Player not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/player/collect": {
      "post": {
        "tags": [
          "Player"
        ],
        "summary": "Collect accumulated idle time",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Collected player state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerCollectResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/player/daily-reward/collect": {
      "post": {
        "tags": [
          "Player"
        ],
        "summary": "Collect daily reward",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated player state after reward",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerState"
                }
              }
            }
          },
          "400": {
            "description": "Reward unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}